ai/routing/requirements.txt
littlehorse-client==0.13.2
typing-extensions==4.13.2
langchain-openai==0.3.14
langchain==0.3.23
dotenv==0.9.9